변경: HeaderPage 코드 스타일 수정 코드 스타일 관련 component 명, component 선언 방식이 중점
