+++ src/app/App.js
@@ -1,7 +1,13 @@
 import React from "react";
+import Header from "../common/components/Header";
 
 function App() {
-  return <div>PPTHUB</div>;
+  return (
+    <div>
+      <Header />
+      <div> mainpage </div>
+    </div>
+  );
 }
 
 export default App;
